Added utilities to merge in additional data.
We can now merge in keys to mark filter flags in a primary maf file.  This is done via a mapreduce strategy and can be done in a pipeline with intermediate sorting.
We actually ran this saving the sorted output because this can later be used with sort -m to merge in new marks, saving us some time in redoing the sorts later.
If you have a Hadoop system running you can use this schema with streaming though be aware that this implementation puts all of the data through a single node.

Added validation-merge.py which is used to add validation AND prior data to a maf file.  We add a new column, MERGESOURCE which indicates either PRIMARY for the
data orriginating from the primary file, MERGE if the record was found in the primary file and the secondary validation/prior file(s) and PASTE if the record
was found only in the validation/prior data but not in the primary.  This will be the main method for adding variants into the MAF with validation data.  The
MERGE records will have the Verification_Status Validation_Status and Mutation_Status of the secondary data pasted into those fields in the primary record.  Note
that the paste operation here will take the values of the first validating record so if there is conflicing information the results are undefined.  You must clean
the input data prior to merger with this tool.